Chore/lockfile sync

+++ neobank-lambdas/transaction-service/src/main/java/neobank/InsufficientFundsException.java
@@ -1,18 +1,19 @@
 package neobank;
 
+import java.math.BigDecimal;
 
 public class InsufficientFundsException extends Exception {
 
-    private final double currentBalance;
-    private final double requiredAmount;
+    private final BigDecimal currentBalance;
+    private final BigDecimal requiredAmount;
 
-    public InsufficientFundsException(double currentBalance, double requiredAmount) {
+    public InsufficientFundsException(BigDecimal currentBalance, BigDecimal requiredAmount) {
         super(String.format("Insufficient funds. Balance: %.2f, Required: %.2f",
                 currentBalance, requiredAmount));
         this.currentBalance = currentBalance;
         this.requiredAmount = requiredAmount;
     }
 
-    public double getCurrentBalance() { return currentBalance; }
-    public double getRequiredAmount() { return requiredAmount; }
+    public BigDecimal getCurrentBalance() { return currentBalance; }
+    public BigDecimal getRequiredAmount() { return requiredAmount; }
 }
